# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Kapusta api",
        "description": "Documentation Kapusta API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://31.128.46.70",
            "description": "Debug test server"
        },
        {
            "url": "http://localhost",
            "description": "Local backend server"
        }
    ],
    "paths": {
...
        "/api/v1/projects": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "Получить список проектов пользователя",
                "operationId": "4581dcc89729ced0a5c29d8132aaaee0",
                "responses": {
                    "200": {
                        "description": "Список проектов",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ProjectList"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Пользователь не авторизован",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Неверный запрос",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Invalid request"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "auth_api": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Projects"
                ],
                "summary": "Создать новый проект",
                "operationId": "7afee26aa9047c66217855bd05d3db0b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProjectRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Проект успешно создан",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProjectDetail"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Ошибка валидации данных",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Validation error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Internal server error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "auth_api": []
                    }
                ]
            }
        },
        "/api/v1/projects/{projectId}": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "Получить информацию о проекте",
                "operationId": "d15a8ad2c684adf95377877800f8e713",
                "parameters": [
                    {
                        "name": "projectId",
                        "in": "path",
                        "description": "Id проекта",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация о проекте",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProjectDetail"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Проект не найден",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Project not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "auth_api": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Projects"
                ],
                "summary": "Удалить проект",
                "operationId": "44986be29c394e143d35e39f29cddec3",
                "parameters": [
                    {
                        "name": "projectId",
                        "in": "path",
                        "description": "Id проекта",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Проект успешно удален"
                    },
                    "404": {
                        "description": "Проект не найден",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Project not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "auth_api": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "ExpenceRequest": {
                "description": "Данные для затрат",
                "properties": {
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": "500.00"
                    },
                    "date": {
                        "type": "string",
                        "format": "string",
                        "example": "1999-01-01"
                    },
                    "comment": {
                        "type": "string",
                        "format": "string",
                        "example": "50",
                        "nullable": true
                    },
                    "projectId": {
                        "type": "integer",
                        "format": "integer",
                        "example": "50",
                        "nullable": true
                    },
                    "categoryId": {
                        "type": "integer",
                        "format": "integer",
                        "example": "50"
                    },
                    "accountId": {
                        "type": "integer",
                        "format": "integer",
                        "example": "50"
                    }
                },
                "type": "object"
            },
            "ProjectRequest": {
                "description": "Данные для создания проекта",
                "properties": {
                    "budget": {
                        "type": "number",
                        "format": "float",
                        "example": "1000.00"
                    },
                    "name": {
                        "type": "string",
                        "format": "string",
                        "example": "Car"
                    },
                    "oldExpenses": {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "example": "{1,2,3}"
                    },
                    "newExpenses": {
                        "type": "array",
                        "format": "array",
                        "items": {
                            "$ref": "#/components/schemas/ExpenceRequest"
                        },
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "Category": {
                "description": "Информация о категории",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int",
                        "example": "1"
                    },
                    "name": {
                        "type": "string",
                        "format": "string",
                        "example": "food"
                    },
                    "purchaseCount": {
                        "type": "integer",
                        "format": "integer",
                        "example": "15"
                    },
                    "icon": {
                        "type": "string",
                        "format": "string",
                        "example": "https://example.com/path/to/file.jpg"
                    },
                    "totalExpenses": {
                        "type": "number",
                        "format": "float",
                        "example": "150.00"
                    },
                    "personOfBudget": {
                        "type": "integer",
                        "format": "integer",
                        "example": "20"
                    }
                },
                "type": "object"
            },
...
            "ProjectLimits": {
                "description": "Информация о лимитах",
                "properties": {
                    "spent": {
                        "type": "number",
                        "format": "float",
                        "example": "500.00"
                    },
                    "limit": {
                        "type": "number",
                        "format": "float",
                        "example": "1000.00"
                    },
                    "percent": {
                        "type": "integer",
                        "format": "integer",
                        "example": "50"
                    },
                    "unit": {
                        "format": "string",
                        "enum": [
                            "$",
                            "€",
                            "£",
                            "₽"
                        ],
                        "example": "$"
                    }
                },
                "type": "object"
            },
            "ProjectDetail": {
                "description": "Детальная информация проекта",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int",
                        "example": "1"
                    },
                    "totalBalance": {
                        "type": "number",
                        "format": "float",
                        "example": "1000.00"
                    },
                    "name": {
                        "type": "string",
                        "format": "string",
                        "example": "Car"
                    },
                    "expenses": {
                        "type": "number",
                        "format": "float",
                        "example": "60000.00"
                    },
                    "limits": {
                        "$ref": "#/components/schemas/ProjectLimits"
                    },
                    "categories": {
                        "type": "array",
                        "format": "array",
                        "items": {
                            "$ref": "#/components/schemas/Category"
                        }
                    },
                    "incomes": {
                        "type": "array",
                        "format": "array",
                        "items": {
                            "type": "string"
                        },
                        "nullable": true
                    },
                    "unit": {
                        "format": "string",
                        "enum": [
                            "$",
                            "€",
                            "£",
                            "₽"
                        ],
                        "example": "$"
                    }
                },
                "type": "object"
            },
            "ProjectList": {
                "description": "Список проектов",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int",
                        "example": "1"
                    },
                    "name": {
                        "type": "string",
                        "format": "string",
                        "example": "Car"
                    },
                    "spent": {
                        "type": "number",
                        "format": "float",
                        "example": "600.00"
                    },
                    "budget": {
                        "type": "number",
                        "format": "float",
                        "example": "60000.00"
                    },
                    "percent": {
                        "type": "integer",
                        "format": "int",
                        "example": "50"
                    },
                    "unit": {
                        "format": "string",
                        "enum": [
                            "$",
                            "€",
                            "£",
                            "₽"
                        ],
                        "example": "$"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}